Move normalizers to chat-utils
