+++ mlflow/server/js/src/shared/web-shared/model-trace-explorer/ModelTraceExplorer.utils.tsx
@@ -51,7 +51,7 @@ import {
   normalizeLlamaIndexChatResponse,
   normalizeDspyChatInput,
   normalizeDspyChatOutput,
-} from './normalizers';
+} from './chat-utils';
 
 export const getCurrentUser = () => {
   return 'User';
