+++ mlflow/server/js/src/shared/web-shared/model-trace-explorer/chat-utils/openai.ts
